+++ NEWS.md
@@ -6,6 +6,7 @@ Have a question, or suggestion, or wanna get in touch? Join the package [DISCUSS
 
 1. Fixed some bugs that did not create problems [#55](https://github.com/bsvars/bsvars/issues/55)
 2. Got rid of filling by reference in the samplers for the sake of granting the exported **cpp** functions usability [#56](https://github.com/bsvars/bsvars/issues/56)
+3. Sparse mixture and Markov-switching models can now have more than 20 regimes [#57](https://github.com/bsvars/bsvars/issues/57)
 
 # bsvars 2.1.0